files/requirements.txt
# Carbon Footprint Tracker - Python Dependencies
# Install karo: pip install -r requirements.txt

# ML & Data
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2

# Backend
flask==3.0.0
flask-cors==4.0.0
PyJWT==2.8.0

# Utilities
pickle-mixin==1.0.2
